@@ -27,4 +27,4 @@
 		/>
 	</div>
 </ic-section-container>
-<hr class="my-10 md:my-4 w-2/3 block m-auto" />
+<hr class="py-4 my-14 md:my-4 w-2/3 block m-auto border-t border-t-gray-200" />
