@@ -2,9 +2,9 @@
 
 import '@testing-library/jest-dom';
 import Switch from './Switch.svelte';
-import { fireEvent } from '@testing-library/svelte';
 import { hydratedRender as render } from '$test-helpers/render';
 import { screen } from 'shadow-dom-testing-library';
+import userEvent from '@testing-library/user-event';
 
 describe('Switch component', () => {
 	const exampleProps = {
@@ -24,13 +24,17 @@ describe('Switch component', () => {
 	});
 
 	it('dispatches a change event when clicked', async () => {
-		const component = (await render(Switch, exampleProps)).component as Switch;
+		const user = userEvent.setup();
 		const changed = vi.fn();
-		component.$on('change', changed);
-
+		await render(Switch, { ...exampleProps, onchange: changed });
 		const theSwitch = await screen.findByShadowLabelText(exampleProps.label);
-		await fireEvent.click(theSwitch);
 
-		expect(changed).toHaveBeenCalledOnce();
+		// Toggle the switch a few times, make sure it's turning off and on
+		await user.click(theSwitch);
+		expect(changed).toHaveBeenCalledWith({ checked: true });
+		await user.click(theSwitch);
+		expect(changed).toHaveBeenCalledWith({ checked: false });
+		await user.click(theSwitch);
+		expect(changed).toHaveBeenCalledWith({ checked: true });
 	});
 });
