@@ -1,7 +1,7 @@
 <!-- (c) Crown Copyright GCHQ -->
 
 <thead class="p-4 text-left break-all">
-	<tr class="border-b-2">
+	<tr class="border-b-gray-200 border-b-2">
 		<slot />
 	</tr>
 </thead>
