@@ -4,6 +4,6 @@ import type { LayoutLoad } from './$types';
 
 export const load = (({ url }) => {
 	return {
-		iri: decodeURIComponent(url.searchParams.get('iri') || '')
+		iri: url.searchParams.get('iri') || ''
 	};
 }) satisfies LayoutLoad;
